Refactor the dependencies after the renaming
