+++ common/logging/logger/format_test.go
@@ -3,8 +3,7 @@ package logger_test
 import (
 	"testing"
 
-	"github.com/kyma-project/kyma/common/logger/logger"
-
+	"github.com/kyma-project/kyma/common/logging/logger"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
