+++ common/logging/logger/logger_test.go
@@ -8,13 +8,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/kyma-project/kyma/common/logger/logger"
-
+	"github.com/kyma-project/kyma/common/logging/logger"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
-
-	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zaptest/observer"
 )
