+++ components/application-connectivity-validator/cmd/applicationconnectivityvalidator/applicationconnectivityvalidator.go
@@ -7,10 +7,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/kyma-project/kyma/common/logger/tracing"
+	"github.com/kyma-project/kyma/common/logging/logger"
+	"github.com/kyma-project/kyma/common/logging/tracing"
 	"github.com/kyma-project/kyma/components/application-connectivity-validator/internal/controller"
-
-	logger "github.com/kyma-project/kyma/common/logger/logger"
 	"github.com/kyma-project/kyma/components/application-connectivity-validator/internal/externalapi"
 	"github.com/kyma-project/kyma/components/application-connectivity-validator/internal/validationproxy"
 	"github.com/patrickmn/go-cache"
