+++ components/application-connectivity-validator/go.mod
@@ -15,7 +15,7 @@ require (
 	github.com/kyma-project/kyma/components/application-operator v0.0.0-20200902071617-83c683010f30
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/patrickmn/go-cache v2.1.0+incompatible
-	github.com/stretchr/testify v1.6.1
+	github.com/stretchr/testify v1.7.0
 	github.com/vrischmann/envconfig v1.1.0
 	go.uber.org/zap v1.16.0
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d // indirect
@@ -33,7 +33,7 @@ require (
 replace (
 	github.com/containerd/containerd => github.com/containerd/containerd v1.4.1
 	github.com/docker/docker => github.com/docker/docker v17.12.0-ce-rc1.0.20200916142827-bd33bbf0497b+incompatible
-	github.com/kyma-project/kyma/common => github.com/franpog859/kyma/common v0.0.0-20210121091704-7ae5af7db5e8
+	github.com/kyma-project/kyma/common => github.com/franpog859/kyma/common v0.0.0-20210125130648-d84b0b196bf9
 	github.com/opencontainers/runc => github.com/opencontainers/runc v1.0.0-rc92
 	golang.org/x/crypto => golang.org/x/crypto v0.0.0-20200604202706-70a84ac30bf9
 	golang.org/x/text => golang.org/x/text v0.3.3
